requirements.txt
streamlit>=1.28.0
pillow-heif>=1.0.0
Pillow>=10.0.0